test: fornisce un localStorage funzionante nei test jsdom su Node 25+
Su Node >=25 il Web Storage built-in e attivo di default: senza
--localstorage-file il global `localStorage` vale `undefined` e, poiche
"esiste" gia, l'ambiente jsdom di vitest non ne installa uno funzionante
sopra. I beforeEach di ThemeProvider e NotificationPrompt chiamavano
`localStorage.clear()` e fallivano con TypeError (9 test). Causa upstream
di Node, non risolta da vitest (issue #8757).

- src/test/setup.ts: installa un polyfill Map-backed come global via
  Object.defineProperty (robusto a qualsiasi shape del built-in di Node)
- src/test/localStorage.ts: helper condiviso makeLocalStorage estratto da
  haptics.test.ts (rimossa la duplicazione)
- vitest.config.ts: registra il setupFiles e disabilita il Web Storage
  sperimentale di Node (execArgv --no-experimental-webstorage), che
  neutralizza la causa radice e silenzia l'ExperimentalWarning

npm test: da 2 file / 9 test falliti a 31 file / 156 test verdi.
Nessuna logica dei componenti toccata, solo infrastruttura di test.

Closes #40
